refactor(chamados): mover models de dominio do chamado para app chamados
